Prevent encode from having to load whole input file into memory ...
THIS IS A BREAKING CHANGE. Bump version to 0.2.0, accordingly.

Now the encode script will only read up to 100MB (configurable with
--chunk-max-bytes) at a time, and compress/encrypt that. Before it would
read the entire file into memory.

Since the decode script now has to decrypt one chunk at a time instead
of all the data at once, this is a breaking change.

The decode script has not yet been modified to not load the entire file
into memory.

No longer are chunks shuffled, as doing so requires us to have them all
to begin with. They are now ordered, with crypt info, then data, then
index.